move array related freelist functions into own file
