+++ freelist.go
@@ -82,11 +82,6 @@ func (f *freelist) count() int {
 	return f.free_count() + f.pending_count()
 }
 
-// arrayFreeCount returns count of free pages(array version)
-func (f *freelist) arrayFreeCount() int {
-	return len(f.ids)
-}
-
 // pending_count returns count of pending pages
 func (f *freelist) pending_count() int {
 	var count int
@@ -105,50 +100,6 @@ func (f *freelist) copyall(dst []common.Pgid) {
 	}
 	sort.Sort(m)
 	common.Mergepgids(dst, f.getFreePageIDs(), m)
-}
-
-// arrayAllocate returns the starting page id of a contiguous list of pages of a given size.
-// If a contiguous block cannot be found then 0 is returned.
-func (f *freelist) arrayAllocate(txid common.Txid, n int) common.Pgid {
-	if len(f.ids) == 0 {
-		return 0
-	}
-
-	var initial, previd common.Pgid
-	for i, id := range f.ids {
-		if id <= 1 {
-			panic(fmt.Sprintf("invalid page allocation: %d", id))
-		}
-
-		// Reset initial page if this is not contiguous.
-		if previd == 0 || id-previd != 1 {
-			initial = id
-		}
-
-		// If we found a contiguous block then remove it and return it.
-		if (id-initial)+1 == common.Pgid(n) {
-			// If we're allocating off the beginning then take the fast path
-			// and just adjust the existing slice. This will use extra memory
-			// temporarily but the append() in free() will realloc the slice
-			// as is necessary.
-			if (i + 1) == n {
-				f.ids = f.ids[i+1:]
-			} else {
-				copy(f.ids[i-n+1:], f.ids[i+1:])
-				f.ids = f.ids[:len(f.ids)-n]
-			}
-
-			// Remove from the free cache.
-			for i := common.Pgid(0); i < common.Pgid(n); i++ {
-				delete(f.cache, initial+i)
-			}
-			f.allocs[initial] = txid
-			return initial
-		}
-
-		previd = id
-	}
-	return 0
 }
 
 // free releases a page and its overflow for a given transaction id.
@@ -286,16 +237,6 @@ func (f *freelist) read(p *common.Page) {
 	}
 }
 
-// arrayReadIDs initializes the freelist from a given list of ids.
-func (f *freelist) arrayReadIDs(ids []common.Pgid) {
-	f.ids = ids
-	f.reindex()
-}
-
-func (f *freelist) arrayGetFreePageIDs() []common.Pgid {
-	return f.ids
-}
-
 // write writes the page ids onto a freelist page. All free and pending ids are
 // saved to disk since in the event of a program crash, all pending ids will
 // become free.
@@ -384,35 +325,4 @@ func (f *freelist) reindex() {
 			f.cache[pendingID] = struct{}{}
 		}
 	}
-}
-
-// arrayMergeSpans try to merge list of pages(represented by pgids) with existing spans but using array
-func (f *freelist) arrayMergeSpans(ids common.Pgids) {
-	sort.Sort(ids)
-	common.Verify(func() {
-		idsIdx := make(map[common.Pgid]struct{})
-		for _, id := range f.ids {
-			// The existing f.ids shouldn't have duplicated free ID.
-			if _, ok := idsIdx[id]; ok {
-				panic(fmt.Sprintf("detected duplicated free page ID: %d in existing f.ids: %v", id, f.ids))
-			}
-			idsIdx[id] = struct{}{}
-		}
-
-		prev := common.Pgid(0)
-		for _, id := range ids {
-			// The ids shouldn't have duplicated free ID. Note page 0 and 1
-			// are reserved for meta pages, so they can never be free page IDs.
-			if prev == id {
-				panic(fmt.Sprintf("detected duplicated free ID: %d in ids: %v", id, ids))
-			}
-			prev = id
-
-			// The ids shouldn't have any overlap with the existing f.ids.
-			if _, ok := idsIdx[id]; ok {
-				panic(fmt.Sprintf("detected overlapped free page ID: %d between ids: %v and existing f.ids: %v", id, ids, f.ids))
-			}
-		}
-	})
-	f.ids = common.Pgids(f.ids).Merge(ids)
 }
